Commit message
**Refactor BusyBox integration and improve build scripts**

Simplified and modularized build scripts by revising BusyBox handling logic. Removed unnecessary argument checks, improved quoting, and added dynamic functionality for better readability and error handling. Streamlined Maven configurations for consistent behavior.

+++ starship-init/init/scripts/install.sh
@@ -5,36 +5,26 @@
 # Licensed under GPL2, GPL3 and Apache 2
 #
 
-function mount_rootfs() {
-  echo "Locating free loop device."
-  sudo mkdir -p "/mnt/rootfs"
-  LOOPDEV=$(sudo losetup -fP --show "../../buildroot/buildroot/output/images/rootfs.ext4")
-  sudo mount -o loop "../../buildroot/buildroot/output/images/rootfs.ext4" "/mnt/rootfs"
-  echo "$LOOPDEV @ /mnt/rootfs"
-}
+# Source the centralized library
+source "../../scripts/fs_library.sh"
 
+INIT_VERSION="1.0.0"
 function make_dirs() {
-  echo "Creating directories."
+  log "Creating directories."
   sudo mkdir -p "/mnt/rootfs/var/lib/starship/" "/mnt/rootfs/etc/starship/config.d"
 }
 
 function copy_files() {
-  echo "Copy files."
-  sudo cp -v "target/init-1.0.0.jar" "/mnt/rootfs/var/lib/starship/init.jar"
+  log "Copying files."
+  sudo cp -v "target/init-${INIT_VERSION}.jar" "/mnt/rootfs/var/lib/starship/init.jar"
   sudo cp -v "src/main/resources/default-init.groovy" "/mnt/rootfs/etc/starship/config.d/init.groovy"
-#  sudo cp -v "src/main/resources/init.sh" "/mnt/rootfs/init"
-  sudo rm -v "/mnt/rootfs/linuxrc"
-#  sudo chmod +x "/mnt/rootfs/init"
-}
-
-function unmount_rootfs() {
-  echo "Cleanup"
-  sudo sync
-  sudo umount "/mnt/rootfs/"
+  if [ -e "/mnt/rootfs/linuxrc" ]; then
+    sudo rm -v "/mnt/rootfs/linuxrc"
+  fi
 }
 
 function main() {
-  mount_rootfs
+  mount_rootfs "../../buildroot/buildroot/output/images/rootfs.ext4"
   make_dirs
   copy_files
   unmount_rootfs
